refactored DB Connection Form made ValidatedInput
